Developing unit tests for entropy regularization.
An issue in the existing implementation of entropy-regularized VI
is the default policy doesn't influence q-values. I forget the details,
but have a hunch that this is only a constant influence, so may be
ok to leave out. In Todorov 2006, it's in equation 18. In Rubin 2012
it's in Algorithm 1, the rho_s(a) term in the line that updates
Z(s;beta).

Still working through other test cases, like reward that varies
per state, states with varying numbers of actions, and stochastic tasks.